requirements.txt
Pillow
pyTelegramBotApi
flask